# Conflicts:
#	send_email.py

+++ Pids.py
@@ -3,6 +3,7 @@ import sys
 import os
 
 def Pi(Core, cpu_load):
+	big_task = []
 	Process_int = 0
 	Processes = []
 	Processes_D = {}
@@ -12,7 +13,8 @@ def Pi(Core, cpu_load):
 	CPU = "CPU"
 	PID_name = "PID_name"
 	warn = False
-	
+	last_hope = False	
+
 	for pid in pids:
 		Process_int = Process_int + 1
 		print Process_int, " current", Processes
@@ -21,7 +23,7 @@ def Pi(Core, cpu_load):
 			pid_name = p.name()
 			cpu = p.cpu_percent(interval=0.2) 
 			
-			if (cpu > 5 * Core) & (cpu < 85 * Core):
+			if (cpu > 5 * Core) & (cpu < 15 * Core):
 	
 				print pid, " ", cpu," " , pid_name
 				Processes.append(pid)
@@ -32,10 +34,9 @@ def Pi(Core, cpu_load):
 				Processes_D.update(PR1)
 				draw = draw + 1
 
-			elif (cpu <= 85 * Core) & (cpu_load >= 90):
-				continue				
-
-			elif cpu >= (85 * Core):
+		
+				
+			elif (cpu >= 85):
 				print pid, " ", cpu," " , pid_name
 				Processes.append(pid)
 				PID = "PID"+str(draw)
@@ -49,7 +50,7 @@ def Pi(Core, cpu_load):
 				os.system('clear')
 		else:
 			os.system('clear')	
-	if (warn != True):
+	if (warn != True) & (last_hope != True):
 		print "total processes: ", Process_int
 		for i in range(0,draw):
 			try:
@@ -63,7 +64,7 @@ def Pi(Core, cpu_load):
 			except:
 				return "%d | %d percent of CPU | %s" % (Processes_D["PID"+str(i)], Processes_D["CPU"+str(i)], Processes_D["PID_name"+str(i)])
 				
-	else: 			
+	elif (warn == True) & (last_hope != True): 			
 		print "WARNING! WARNING! System will be repaired..."
 		p.terminate()
 		print "terminate..."
@@ -71,8 +72,15 @@ def Pi(Core, cpu_load):
 		print "kill..."
 		warn = False
 		return PID, " ", PID_name, "could be a problem for you, and may could damage your PC! Please keep that in mind."
-def Rescue(P):
-	
+
+	else:
+		Rescue(big_task)
+"""def Rescue(big_task):
+	for i in reversed(big_task)[:2]:
+		cpu, p in i		
+		p.terminate()
+		p.kill()
+	"""
